feat: better interface
The new interface enables loading the value for Cr or Ci (each taking 4 clock cycles), instead of having to load both every time (taking 8 clock cycles). Also, the user can decide when to load the values (instead of having to load them right before starting the calculation), so the user can start loading the next set of values while the calculation is going.